middleware/requirements.txt
fastapi
uvicorn[standard]
fastmcp
graphiti-core
tree-sitter
tree-sitter-python
tree-sitter-javascript
tree-sitter-languages
openai
pydantic-settings
python-dotenv
aiosqlite
